# identy-service/docs/swagger.yaml
basePath: /
definitions:
...
  http.classifierReqBody:
    properties:
      aphasiaType:
        type: string
      difficulty:
        description: "@Description\tClassifier validation payload\n\t@Example\t\t{\"difficulty\":\"hard\",\"aphasiaType\":\"none\"}"
        type: string
    required:
    - aphasiaType
    - difficulty
    type: object
...
  http.loginReq:
    properties:
      email:
        description: "@Description\tLogin payload\n\t@Example\t\t{\"email\":\"user@example.com\",\"password\":\"secret\",\"fingerprint\":\"fp\"}"
        type: string
      fingerprint:
        type: string
      password:
        type: string
    required:
    - email
    - fingerprint
    - password
    type: object
  http.passwordResetConfirmReq:
    properties:
      code:
        type: string
      email:
        description: "@Description\tPassword reset confirmation payload\n\t@Example\t\t{\"email\":\"user@example.com\",\"code\":\"123456\",\"newPassword\":\"StrongPass123\"}"
        type: string
      newPassword:
        maxLength: 72
        minLength: 8
        type: string
    required:
    - code
    - email
    - newPassword
    type: object
  http.passwordResetReq:
    properties:
      email:
        description: "@Description\tEmail for password reset\n\t@Example\t\tuser@example.com"
        type: string
    required:
    - email
    type: object
...
  http.registerReq:
    properties:
      classifier:
        $ref: '#/definitions/http.classifierReqBody'
      email:
        description: "@Description\tRegistration payload\n\t@Example\t\t{\"email\":\"user@example.com\",\"verifyCode\":\"123456\",\"password\":\"StrongPass123\",\"role\":\"user\",\"fingerprint\":\"fp\",\"classifier\":{\"difficulty\":\"hard\",\"aphasiaType\":\"none\"}}"
        type: string
      fingerprint:
        type: string
      password:
        maxLength: 72
        minLength: 8
        type: string
      role:
        type: string
      verifyCode:
        type: string
    required:
    - classifier
    - email
    - fingerprint
    - password
    - role
    - verifyCode
    type: object
  http.registerVerifyReq:
    properties:
      email:
        description: "@Description\tEmail для отправки verify-кода\n\t@Example\t\tuser@example.com"
        type: string
    required:
    - email
    type: object
info:
  contact: {}
  description: HTTP API для управления идентификацией и доступом SayMo.
  title: SayMo Identity Service API
  version: "1.0"
paths:
...
  /iam/auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.loginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login with email and password
      tags:
      - Auth
  /iam/auth/logout:
    post:
      description: 'Требует Authorization: Bearer <JWT>.'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Logout and terminate session
      tags:
      - Auth
  /iam/auth/oauth/google:
    get:
      description: Начало процесса аутентификации через Google
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Initiate Google OAuth flow
      tags:
      - Auth
  /iam/auth/oauth/google/callback:
    get:
      consumes:
      - application/json
      description: Завершение процесса аутентификации через Google (возврат access_token
        и session_id)
      parameters:
      - description: Authorization code
        in: query
        name: code
        required: true
        type: string
      - description: OAuth state
        in: query
        name: state
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Google OAuth callback
      tags:
      - Auth
  /iam/auth/password-reset:
    post:
      consumes:
      - application/json
      description: 'Anti-enumeration: идентичный ответ независимо от существования
        email'
      parameters:
      - description: Email для отправки кода reset
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.passwordResetReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Request password reset code
      tags:
      - Auth
  /iam/auth/password-reset/confirm:
    post:
      consumes:
      - application/json
      description: Подтверждает OTP и меняет пароль после сброса.
      parameters:
      - description: Email, код сброса и новый пароль
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.passwordResetConfirmReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Confirm password reset code
      tags:
      - Auth
  /iam/auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Registration payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.registerReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new account
      tags:
      - Auth
  /iam/auth/register/verify:
    post:
      consumes:
      - application/json
      description: 'Anti-enumeration: всегда возвращает 200 независимо от существования
        email.'
      parameters:
      - description: Email для отправки verify-кода
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.registerVerifyReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Request email verification code for registration
      tags:
      - Auth
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'JWT access token. Формат: Bearer {token}.'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"